tools/aspect_ratio_requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.12.1=h00a0c3c_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.11.4=haa95532_0
cairo=1.18.4=he9e932c_0
dav1d=1.2.1=h2bbff1b_0
expat=2.7.3=h9214b88_0
fontconfig=2.15.0=hd211d86_0
freeglut=3.4.0=h8a1e904_1
freetype=2.13.3=h0620614_0
fribidi=1.0.10=h62dcd97_0
graphite2=1.3.14=hd77b12b_1
harfbuzz=10.2.0=he2f9f60_1
icu=73.1=h6c2663c_0
jpeg=9f=ha349fce_0
lcms2=2.16=h62be587_1
lerc=4.0.0=h5da7b33_0
libavif=1.1.1=h952b276_1
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libglib=2.84.4=hfaec014_0
libiconv=1.16=h2bbff1b_3
libopenjpeg=2.5.4=h02ab6af_1
libpng=1.6.50=h46444df_0
libtiff=4.7.0=h404307b_0
libwebp-base=1.3.2=h3d04722_1
libxml2=2.13.9=h6201b9f_0
libzlib=1.3.1=h02ab6af_0
lz4-c=1.9.4=h2bbff1b_1
openssl=3.0.18=h543e019_0
pcre2=10.46=h5740b90_0
pillow=12.0.0=py312h4212202_1
pip=25.3=pyhc872135_0
pixman=0.46.4=h4043f72_0
python=3.12.12=h716150d_1
setuptools=80.9.0=py312haa95532_0
sqlite=3.51.0=hda9a48d_0
tk=8.6.15=hf199647_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
vc=14.42=haa95532_5
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
wheel=0.45.1=py312haa95532_0
xz=5.6.4=h4754444_1
zlib=1.3.1=h02ab6af_0
zstd=1.5.7=h56299aa_0
